pyproject.toml
[project]
name = "clinicedc"
version = "2.0.41"
description = "A clinical trials data management framework built on Django"
readme = {file = "README.rst", content-type = "text/x-rst"}
keywords = ["django", "clinicedc", "edc", "clinical trials", "research", "data management", "esource"]
authors = [
    { name = "Erik van Widenfelt", email = "ew2789@gmail.com" },
    { name = "Jonathan Willitts", email = "j.willitts@ucl.ac.uk" }
]
maintainers = [
    { name = "Erik van Widenfelt", email = "ew2789@gmail.com" },
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 5.2",
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
license = "GPL-3.0-or-later"
license-files = ["LICEN[CS]E*"]
requires-python = '>=3.12,<3.14'
dependencies = [
    "argon2-cffi>=25.1.0",
    "arrow>=1.3.0",
    "beautifulsoup4>=4.13.5",
    "boto3>=1.39.16",
    "celery>=5.5.3",
    "clinicedc-utils>=0.4.0",
    "clinicedc-constants>=1.1.0",
    "cryptography>=45.0.7",
    "django>=5.2.4",
    "django-admin-rangefilter>=0.13.3",
    "django-audit-fields>=2.0.3",
    "django-crypto-fields>=1.1.2",
    "django-db-views>=0.1.11",
    "django-defender>=0.9.8",
    "django-environ>=0.12.0",
    "django-logentry-admin>=1.1.0",
    "django-multisite2>=2.1.0",
    "django-pandas>=0.6.7",
    "django-pylabels>=1.0.1",
    "django-redis>=6.0.0",
    "django-revision>=2.0.1",
    "django-sequences>=3.0",
    "django-simple-history>=3.10.1",
    "django-storages>=1.14.6",
    "djangorestframework>=3.16.0",
    "fontawesomefree>=6.6.0",
    "gunicorn>=23.0.0",
    "inflect>=7.5.0",
    "mempass>=0.1.2",
    "mysqlclient>=2.2.7",
    "psycopg2>=2.9.10",
    "pycups>=2.0.4",
    "pypdf>=5.9.0",
    "python-dateutil>=2.9.0.post0",
    "python-memcached>=1.62",
    "reportlab>=4.4.3",
    "sqlalchemy>=2.0.42",
    "sqlglot>=27.8.0",
    "tqdm>=4.67.1",
    "twilio>=9.7.0",
]

[project.urls]
Homepage = "https://github.com/clinicedc/clinicedc"
Documentation = "https://github.com/clinicedc/clinicedc/docs"
Repository = "https://github.com/clinicedc/clinicedc.git"
Changelog = "https://github.com/clinicedc/clinicedc/blob/main/CHANGES"

[build-system]
requires = ["uv_build>=0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-root = "src"
module-name = [
    "clinicedc",
    "edc_action_item",
    "edc_adherence",
    "edc_adverse_event",
    "edc_analytics",
    "edc_appconfig",
    "edc_appointment",
    "edc_auth",
    "edc_consent",
    "edc_constants",
    "edc_crf",
    "edc_dashboard",
    "edc_data_manager",
    "edc_device",
    "edc_document_status",
    "edc_dx",
    "edc_dx_review",
    "edc_egfr",
    "edc_export",
    "edc_facility",
    "edc_fieldsets",
    "edc_form_describer",
    "edc_form_label",
    "edc_form_runners",
    "edc_form_validators",
    "edc_glucose",
    "edc_identifier",
    "edc_lab",
    "edc_lab_dashboard",
    "edc_lab_panel",
    "edc_lab_results",
    "edc_label",
    "edc_list_data",
    "edc_listboard",
    "edc_locator",
    "edc_ltfu",
    "edc_metadata",
    "edc_model",
    "edc_model_admin",
    "edc_model_fields",
    "edc_model_form",
    "edc_model_to_dataframe",
    "edc_navbar",
    "edc_notification",
    "edc_offstudy",
    "edc_pdf_reports",
    "edc_pdutils",
    "edc_pharmacy",
    "edc_prn",
    "edc_protocol",
    "edc_protocol_incident",
    "edc_pylabels",
    "edc_qareports",
    "edc_randomization",
    "edc_refusal",
    "edc_registration",
    "edc_reportable",
    "edc_review_dashboard",
    "edc_screening",
    "edc_search",
    "edc_sites",
    "edc_subject_dashboard",
    "edc_timepoint",
    "edc_transfer",
    "edc_unblinding",
    "edc_utils",
    "edc_view_utils",
    "edc_visit_schedule",
    "edc_visit_tracking",
    "edc_vitals",
]
source-include = ["edc_*/templates/**", "edc_*/locale/**", "locale/**"]
source-exclude = ["edc_*/tests/**", "edc_*/notebooks/**", "tests/**", "bin"]
namespace = false

[tool.ruff]
line-length = 95
indent-width = 4
target-version = "py312"
extend-exclude = ["migrations", ".tox", ".venv"]

[tool.ruff.lint]
select = [
    "A",
    "ARG",
    "B",
    "C90",
    "COM",
    "DJ",
    "DTZ",
    "E",
    "F",
    "FA",
    "FURB",
    "I",
    "INT",
    "N",
    "PD",
    "PERF",
    "PL",
    "PLR",
    "PTH",
    "RET",
    "RUF",
    "S",
    "SIM",
    "T20",
    "TRY",
    "UP",
    "W",
    # "FIX",
]
ignore = [
    "DJ008",
    "DJ007",
    "DJ006",
    "COM812",
    "PLC0206",
    "TRY003", # temporarily
    "PLR0913", # temporarily
    "ARG001", # temporarily
    "DJ012", # temporarily
]

[tool.ruff.lint.per-file-ignores]
    "__init__.py" = ["F401"]

[tool.ruff.lint.mccabe]
max-complexity = 12

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.coverage.run]
parallel = false
branch = true
source = ["src"]

[tool.coverage.paths]
source = ["src"]

[tool.coverage.report]
show_missing = true
skip_covered = true
omit = ["requirements.txt", "*/migrations/*"]
exclude_lines = [
  "pragma: no cover",
  "if TYPE_CHECKING:",
]

[tool.rstcheck]
ignore_directives = ["toctree"]

[dependency-groups]
docs = [
    "furo>=2025.7.19",
    "sphinx>=8.2.3",
    "sphinx-design>=0.6.1",
    "sphinx-inline-tabs>=2023.4.21",
]
lint = [
    "ruff>=0.13.2",
]

test = [
    "arrow>=1.3.0",
    "clinicedc-tests>=0.1.5",
    "coverage>=7.10.1",
    "django-audit-fields>=2.0.0",
    "django-webtest>=1.9.13",
    "edc-test-settings>=2.0.0",
    "faker>=37.4.2",
    "model-bakery>=1.20.5",
    "time-machine>=2.16.0",
    "tox-uv>=1.26.2",
]

notebook = [
    "dj-notebook>=0.7.0",
    "great-tables>=0.18.0",
    "matplotlib>=3.10.3",
    "matplotlib-venn>=1.1.2",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "pdfkit>=1.0.0",
    "scikit-learn>=1.7.1",
    "scipy>=1.16.0",
    "selenium>=4.34.2",
    "tabulate>=0.9.0",
]

develop = [
    "clinicedc",
    "clinicedc-tests",
    'argon2-cffi',
    'python-memcached',
    'django-environ',
    'django-storages',
    'fontawesomefree',
    "django-audit-fields",
    'django-crypto-fields',
    'django-revision',
    'django-multisite2',
    'django-simple-history',
    'django-pandas',
    'django-logentry-admin',
    'django-admin-rangefilter',
    'django-db-views',
    "model-bakery>=1.20.5",
    "faker>=37.5.3",
    "time-machine>=2.17.0",
    "ipython>=9.5.0",
]

[tool.uv.sources]
clinicedc = { workspace = true }
clinicedc-tests = { path = "../clinicedc-tests", editable = true }
clinicedc-utils = { path = "../clinicedc-utils", editable = true }
django-audit-fields = { path = "../django-audit-fields", editable = true }
django-crypto-fields = { path = "../django-crypto-fields", editable = true }
django-revision = { path = "../django-revision", editable = true }
# django-multisite2 = { path = "../django-multisite2", editable = true }
edc-test-settings = { path = "../edc-test-settings", editable = true }
